+++ NEWS.md
@@ -1,3 +1,6 @@
+# trace.package 0.1.1
+* Address CRAN submission comments
+
 # trace.package 0.1.0
 
 * Initial CRAN submission.